requirements.txt
groq
streamlit
python-dotenv
pyttsx3
TextBlob
